% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_indep_threevectors.R
\name{lin_indep_threevectors}
\alias{lin_indep_threevectors}
\title{Checks if three vectors are linearly independent and returns true if they are.}
\usage{
lin_indep_threevectors(x, y, z, verbose = FALSE, na.rm = FALSE)
}
\arguments{
\item{x}{Numeric vector}

\item{y}{Numeric vector}

\item{z}{Numeric vector}

\item{verbose}{if true, it prints the progress in computing the 3x3 minors of the matrix \link{x y z}}

\item{na.rm}{}
}
\value{
True if the three vectors are linearly independent.
}
\description{
Checks if three vectors are linearly independent and returns true if they are.
}
\examples{
lin_indep_threevectors(c(2,1,1),c(3,2,1),c(5,5,6))

}
